% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelTVP_IV.R
\name{panelTVP_IV}
\alias{panelTVP_IV}
\title{Fit a Bayesian IV panel data model with time-varying parameters}
\usage{
panelTVP_IV(formula_stage1 = NULL,
            formula_stage2 = NULL,
            data = NULL,
            id = NULL,
            t = NULL,
            prior.reg_stage1 = list(),
            prior.reg_stage2 = list(),
            prior.var_stage2 = list(),
            prior.load_stage2 = list(),
            prior.rho = list(),
            mcmc.opt = list(),
            HPD.coverage = 0.95,
            posterior.predictive.matrix = FALSE,
            random.effects = TRUE,
            progress.bar = FALSE)
}
\arguments{
\item{formula_stage1}{formula object for the first-stage Probit model (no default)}

\item{formula_stage2}{formula object for the second-stage Gaussian model (no default)}

\item{data}{a data frame that contains the variables of the formula arguments
(no default)}

\item{id}{a vector with length equal to the number of observations in the dataset, i.e.,
this is the 'subject' variable in the dataset (no default)}

\item{t}{a vector with length equal to the number of observations in the dataset, i.e.,
this is the 'time' variable in the dataset (no default)}

\item{prior.reg_stage1}{a list of arguments for estimating the parameters of the regression
part of the first-stage Probit model. Note that this stage does not contain
random effects. The prior arguments are the ones of \code{\link[=panelTVP]{panelTVP()}}}

\item{prior.reg_stage2}{a list of arguments for estimating the parameters of the regression
part of the second-stage Gaussian model. The prior arguments are the ones of \code{\link[=panelTVP]{panelTVP()}}.}

\item{prior.var_stage2}{a list of arguments for estimating the error variance
of the second-stage Gaussian model using 2D-Slice sampling. The arguments are:
\itemize{
\item \code{alpha.sigma}: shape parameter of Inverse-Gamma prior
\item \code{beta.sigma}: rate parameter of Inverse-Gamma prior
\item \code{expansion.steps}: number of steps in stepping-out phase
\item \code{width}: width of the slice interval (on log-scale of error variance)
}}

\item{prior.load_stage2}{a list of arguments for estimating the parameters of the factor
part of the second-stage Gaussian model. The prior arguments are the ones of \code{\link[=panelTVP]{panelTVP()}}}

\item{prior.rho}{a list of arguments for estimating the correlation of errors,
i.e. the degree of endogeneity using 2D-Slice sampling. The arguments are:
\itemize{
\item \code{mean.rho}: mean of Normal prior on Fisher-Z scale of \code{rho}
\item \code{sd.rho}: standard deviation of Normal prior on Fisher-Z scale of \code{rho}
\item \code{expansion.steps}: number of steps in stepping-out phase
\item \code{width}: width of the slice interval (on Fisher-Z scale of \code{rho})
}}

\item{mcmc.opt}{a list containing information on the overall sampler.
The arguments are:
\itemize{
\item \code{chain.length}: the length of the Markov Chain (i.e., total number
of draws)
\item \code{burnin}: the burn-in period
\item \code{thin}: the thinning factor
\item \code{asis}: if set to TRUE, an ancillarity sufficiency interweaving
step is added for increasing the sampling efficiency of the regression
effects and factor loadings
}}

\item{HPD.coverage}{coverage probability of highest posterior density intervals
(default yields 95 percent coverage)}

\item{posterior.predictive.matrix}{if TRUE (= default) the posterior predictive distribution
based on the training data is computed and returned as a list object by the function.
Setting it to FALSE is usually only done for saving memory}

\item{random.effects}{if TRUE (= default) a factor model is included for estimating
random effects in the second-stage of the IV model,
if FALSE the model does not contain random effects and, consequently,
priors on the parameters of the factor model are ignored}

\item{progress.bar}{if TRUE a progress bar is displayed, if FALSE the progress bar is omitted}
}
\value{
The function returns an object of class \code{panelTVP.IV} and
contains a list of the following elements:
\describe{
\item{learning.settings_stage1}{information on which parameters have been learned
for the first stage of the model}
\item{learning.settings_stage2}{information on which parameters have been learned
for the second stage of the model}
\item{variable.codes_stage1}{information on which covariate is associated with which parameter
for the first stage of the model}
\item{variable.codes_stage2}{information on which covariate is associated with which parameter
for the second stage of the model}
\item{data}{the data used for fitting the model and additional context information
derived from the data}
\item{Y}{the \eqn{Tn \times M} response data matrix of every iteration of the chain,
i.e., this matrix is only included in the output when missing response data
were present and imputed via data augmentation. Each row of \code{Y} contains
the sampled values of a specific observation. For observed data, the corresponding
rows are essentially replicates of the same value. For missing data, the
corresponding rows contain the imputed values of every iteration.}
\item{mcmc_stage1}{Markov Chains for every parameter except for the factor scores
(to save memory) for the first stage of the model}
\item{mcmc_stage2}{Markov Chains for every parameter except for the factor scores
(to save memory) for the second stage of the model}
\item{mcmc_rho}{a vector of dimension M containing the Markov Chain of the
correlation coefficient}
\item{posterior_stage1}{preliminary summary of posterior results for the
first stage of the model}
\item{posterior_stage2}{preliminary summary of posterior results for the
second stage of the model}
\item{posterior_rho}{preliminary summary of posterior results for the
correlation coefficient}
\item{fmean_stage2}{posterior means of factor scores for the
second stage of the model (only when random effects are estimated)}
\item{model}{the fitted model}
\item{acceptance.rates}{the achieved acceptance rates when using
Metropolis-Hastings for both components of the model}
\item{HPD.coverage}{coverage probability of HPD intervals (based on input)}
\item{runtime}{total runtime of the sampler (measured in seconds)}
\item{Treatment.Variable}{name of the treatment variable of the model}
\item{posterior.predictive}{the \eqn{Tn \times M} matrix that contains the posterior
predictive distribution for each observation (rows) and each MCMC draw (columns)}
\item{mcmc.settings}{details on the MCMC sampler}
}
}
\description{
This function estimates a Bayesian Instrumental Variable Model for identifying
causal treatment effects. The function assumes that the (endogenous) treatment
is binary whereas a Gaussian model is used for the observation equation (second stage).
Similarly to the function \code{\link[=panelTVP]{panelTVP()}}, effect estimates vary over time and
shrinkage priors are used to classify effects to time-varying, constant and
zero. Unobserved heterogeneity is captured by random effects at the second stage
via a factor model. The error variance of the second stage model as well as
the correlation between errors of both stages are sampled jointly via
2D-Slice sampling (Neal, 2003). For this, an Inverse-Gamma prior is used for the variance
whereas a Normal prior is placed on the Fisher-Z scale of the correlation
coefficient. Time-varying effects are modelled with the same priors as in
the non-treatment-effects models in \code{\link[=panelTVP]{panelTVP()}}.
}
\examples{
# Fitting Bayesian IV panel model with time-varying parameters
# NB: To reduce computational effort, we have drastically reduced the length
# of the Markov Chain. You should use a much longer chain in your applications.
sim.iv <- sim_panelTVP_IV(n = 1000,
                          Tmax = 4,
                          beta_stage1 = c(1, 0.7),
                          theta_stage1 = c(0.2, 0.01),
                          beta_stage2 = c(0, 4, 1),
                          theta_stage2 = c(0, 3, 0),
                          lambda_stage2 = 1.3,
                          psi_stage2 = 0.1,
                          beta_D = 2,
                          theta_D = 0.7,
                          rho = 0.1,
                          sigma2 = 1,
                          n.instruments = 1)
res.iv <- panelTVP_IV(formula_stage1 = D ~ X_stage1.Z1,
                      formula_stage2 = y ~ X_stage2.W1 + X_stage2.W2 + D,
                      data = sim.iv$observed,
                      id = sim.iv$observed$id,
                      t = sim.iv$observed$t,
                      prior.rho = list(
                      mean.rho = atanh(0.1), sd.rho = 0.1,
                       expansion.steps = 10, width = 0.1
                      ),
                      mcmc.opt = list(chain.length = 200, burnin = 100, thin = 1, asis = TRUE))
}
\references{
Neal, R.M. (2003). Slice sampling. In: The Annals of Statistics, 31, 705-767.
}
