% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_plot.R
\name{plot.panelTVP.Gaussian}
\alias{plot.panelTVP.Gaussian}
\title{Get plots of time-varying parameters based on a \code{panelTVP.Gaussian} object}
\usage{
\method{plot}{panelTVP.Gaussian}(x, nplots = 4, ...)
}
\arguments{
\item{x}{an object of class \code{panelTVP.Gaussian}}

\item{nplots}{indicates how many plots should be printed on one page}

\item{...}{optional arguments passed to the function (those are ignored)}
}
\description{
This \code{plot} function produces effect plots for the time-varying parameters,
i.e., \eqn{\boldsymbol{\beta}_1,\dots,\boldsymbol{\beta}_T} and
\eqn{\lambda_1,\dots,\lambda_T} separately for each covariate. Plots are produced
using the \code{ggplot2} library.
}
\examples{
# Plot results from an object of class panelTVP.Gaussian
# NB: To reduce computational effort, we have drastically reduced the length
# of the Markov Chain. You should use a much longer chain in your applications.
sim.gaussian <- sim_panelTVP(n = 100,
                             Tmax = 4,
                             beta = c(4,1,0,0),
                             theta = c(1,0.5,0,0),
                             lambda = 1,
                             psi = 0.2,
                             model = "Gaussian",
                             sigma2 = 0.7)
res.gaussian <- panelTVP(y ~ W1 + W2 + W3,
                         data = sim.gaussian$observed,
                         id = sim.gaussian$observed$id,
                         t = sim.gaussian$observed$t,
                         mcmc.opt = list(chain.length = 200, burnin = 100, thin = 1, asis = TRUE),
                         model = "Gaussian")
plot(res.gaussian, nplots = 1)
}
\references{
Wickham, H. (2016). \code{ggplot2}: Elegant Graphics for Data Analysis.
Springer Verlag, New York.
}
\author{
Roman Pfeiler, Helga Wagner
}
