% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_plot.R
\name{plot.panelTVP.IV}
\alias{plot.panelTVP.IV}
\title{Get plots of time-varying parameters based on a \code{panelTVP.IV} object}
\usage{
\method{plot}{panelTVP.IV}(x, component = NULL, nplots = 4, ...)
}
\arguments{
\item{x}{an object of class \code{panelTVP.IV}}

\item{component}{either 'stage1' or 'stage2' to create plots for either the first
stage or the second stage}

\item{nplots}{indicates how many plots should be printed on one page}

\item{...}{optional arguments passed to the function (those are ignored)}
}
\description{
This \code{plot} function produces effect plots for the time-varying parameters,
of an instrumental variable model. Plots are produced
using the \code{ggplot2} library. Note that this
function either plots the effects of the first stage (Probit model) or the
second stage (Gaussian model), where the treatment variable effect is included.
}
\examples{
# Plot results from an object of class panelTVP.IV
# NB: To reduce computational effort, we have drastically reduced the length
# of the Markov Chain. You should use a much longer chain in your applications.
sim.iv <- sim_panelTVP_IV(n = 1000,
                          Tmax = 4,
                          beta_stage1 = c(1, 0.7),
                          theta_stage1 = c(0.2, 0.01),
                          beta_stage2 = c(0, 4, 1),
                          theta_stage2 = c(0, 3, 0),
                          lambda_stage2 = 1.3,
                          psi_stage2 = 0.1,
                          beta_D = 2,
                          theta_D = 0.7,
                          rho = 0.1,
                          sigma2 = 1,
                          n.instruments = 1)
res.iv <- panelTVP_IV(formula_stage1 = D ~ X_stage1.Z1,
                      formula_stage2 = y ~ X_stage2.W1 + X_stage2.W2 + D,
                      data = sim.iv$observed,
                      id = sim.iv$observed$id,
                      t = sim.iv$observed$t,
                      prior.rho = list(
                      mean.rho = atanh(0.1), sd.rho = 0.1,
                       expansion.steps = 10, width = 0.1
                      ),
                      mcmc.opt = list(chain.length = 200, burnin = 100, thin = 1, asis = TRUE))
plot(res.iv, nplots = 1, component = "stage1") # Probit component
plot(res.iv, nplots = 1, component = "stage2") # Gaussian component
}
\references{
Wickham, H. (2016). \code{ggplot2}: Elegant Graphics for Data Analysis.
Springer Verlag, New York.
}
\author{
Roman Pfeiler, Helga Wagner
}
