% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_print.R
\name{print.panelTVP.Gaussian}
\alias{print.panelTVP.Gaussian}
\title{Print basic model information output for a \code{panelTVP.Gaussian} object}
\usage{
\method{print}{panelTVP.Gaussian}(x, ...)
}
\arguments{
\item{x}{an object of class \code{panelTVP.Gaussian}}

\item{...}{optional arguments passed to the function (those are ignored)}
}
\description{
This basic \code{print} method gives a general overview
of the information that is contained in an object of class \code{panelTVP.Gaussian}.
}
\examples{
# Printing object of class panelTVP.Gaussian
# NB: To reduce computational effort, we have drastically reduced the length
# of the Markov Chain. You should use a much longer chain in your applications.
sim.gaussian <- sim_panelTVP(n = 100,
                             Tmax = 4,
                             beta = c(4,1,0,0),
                             theta = c(1,0.5,0,0),
                             lambda = 1,
                             psi = 0.2,
                             model = "Gaussian",
                             sigma2 = 0.7)
res.gaussian <- panelTVP(y ~ W1 + W2 + W3,
                         data = sim.gaussian$observed,
                         mcmc.opt = list(chain.length = 200, burnin = 100, thin = 1, asis = TRUE),
                         model = "Gaussian")
print(res.gaussian)
}
\author{
Roman Pfeiler, Helga Wagner
}
