% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_print.R
\name{print.panelTVP.IV}
\alias{print.panelTVP.IV}
\title{Print basic model information output for a \code{panelTVP.IV} object}
\usage{
\method{print}{panelTVP.IV}(x, ...)
}
\arguments{
\item{x}{an object of class \code{panelTVP.IV}}

\item{...}{optional arguments passed to the function (those are ignored)}
}
\description{
This basic \code{print} method gives a general overview
of the information that is contained in an object of class \code{panelTVP.IV}.
}
\examples{
# Printing object of class panelTVP.IV
# NB: To reduce computational effort, we have drastically reduced the length
# of the Markov Chain. You should use a much longer chain in your applications.

sim.iv <- sim_panelTVP_IV(n = 1000,
                          Tmax = 4,
                          beta_stage1 = c(1, 0.7),
                          theta_stage1 = c(0.2, 0.01),
                          beta_stage2 = c(0, 4, 1),
                          theta_stage2 = c(0, 3, 0),
                          lambda_stage2 = 1.3,
                          psi_stage2 = 0.1,
                          beta_D = 2,
                          theta_D = 0.7,
                          rho = 0.1,
                          sigma2 = 1,
                          n.instruments = 1)
res.iv <- panelTVP_IV(formula_stage1 = D ~ X_stage1.Z1,
                      formula_stage2 = y ~ X_stage2.W1 + X_stage2.W2 + D,
                      data = sim.iv$observed,
                      id = sim.iv$observed$id,
                      t = sim.iv$observed$t,
                      prior.rho = list(
                      mean.rho = atanh(0.1), sd.rho = 0.1,
                       expansion.steps = 10, width = 0.1
                      ),
                      mcmc.opt = list(chain.length = 200, burnin = 100, thin = 1, asis = TRUE))
print(res.iv)
}
\author{
Roman Pfeiler, Helga Wagner
}
