% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_print.R
\name{print.panelTVP.NegBin}
\alias{print.panelTVP.NegBin}
\title{Print basic model information output}
\usage{
\method{print}{panelTVP.NegBin}(x, ...)
}
\arguments{
\item{x}{an object of class \code{panelTVP.NegBin}}

\item{...}{optional arguments passed to the function (those are ignored)}
}
\description{
This basic \code{print} method gives a general overview
of the information that is contained in an object of class \code{panelTVP.NegBin}.
}
\author{
Roman Pfeiler
}
