% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_summary.R
\name{print.summary.panelTVP}
\alias{print.summary.panelTVP}
\title{Print method for summary S3 objects}
\usage{
\method{print}{summary.panelTVP}(x, ...)
}
\arguments{
\item{x}{a summary object}

\item{...}{optioonal arguments passed to the function (those are ignored)}
}
\description{
Print method for summary S3 objects
}
