% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_panelTVP.R
\name{sim_panelTVP}
\alias{sim_panelTVP}
\title{Simulate data from a time-varying parameter panel data model}
\usage{
sim_panelTVP(
  n,
  t,
  beta = NULL,
  theta = NULL,
  lambda = NULL,
  psi = NULL,
  model,
  r = NULL,
  sigma2 = NULL
)
}
\arguments{
\item{n}{number of subjects (scalar)}

\item{t}{number of time points (scalar)}

\item{beta}{starting values of random walk prior for regression effects
(vector of dimension d)}

\item{theta}{standard deviation of random walk prior for regression effects
(vector of dimension d)}

\item{lambda}{starting value of random walk prior for factor loading (scalar)}

\item{psi}{standard deviation of random walk prior for factor loading (scalar)}

\item{model}{either "Gaussian", "Probit", "Logit" or "NegBin"}

\item{r}{dispersion parameter of the Negative Binomial model (scalar, ignored
if response is not simulated from a Negative Binomial regression model)}

\item{sigma2}{homoscedastic error variance of the Gaussian model (scalar,
ignored if response is not simulated from a Gaussian regression model)}
}
\value{
a list with simulated data and parameters
}
\description{
This function simulates balanced panel data, where the distribution of the
response variable is either Gaussian, binary (using a Probit or Logit link)
or Negative Binomial.
}
\examples{
1

}
