% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_panelTVP_IV.R
\name{sim_panelTVP_IV}
\alias{sim_panelTVP_IV}
\title{Simulate data for an instrumental variable model with time-varying parameters
and a binary treatment}
\usage{
sim_panelTVP_IV(
  n,
  Tmax,
  beta_stage1,
  theta_stage1,
  beta_stage2,
  theta_stage2,
  lambda_stage2,
  psi_stage2,
  beta_D,
  theta_D,
  rho,
  sigma2,
  n.instruments = 1
)
}
\arguments{
\item{n}{number of subjects (scalar)}

\item{Tmax}{number of time points / repeated measurements per subject (scalar)}

\item{beta_stage1}{fixed regression effects of first stage (Probit model)
with the first value representing the global intercept (vector of dimension d1)}

\item{theta_stage1}{standard deviations of random walk for regression effects
of first stage (Probit model); larger values yield regression effects
that vary stronger over time (vector of dimension d1)}

\item{beta_stage2}{fixed regression effects of second stage (Gaussian model)
with the first value representing the global intercept (vector of dimension d2)}

\item{theta_stage2}{standard deviations of random walk for regression effects
of second stage (Gaussian model); larger values yield regression effects
that vary stronger over time (vector of dimension d2)}

\item{lambda_stage2}{fixed factor loading in second stage (Gaussian model) (scalar)}

\item{psi_stage2}{standard deviation of random walk for factor loading
in second stage (Gaussian model) (scalar)}

\item{beta_D}{fixed regression effect of treatment variable, which will be
added to the linear predictor of the second stage (scalar)}

\item{theta_D}{standard deviation of random walk for treatment effect (scalar)}

\item{rho}{correlation between errors of both stages controlling the extent of
endogeneity (scalar)}

\item{sigma2}{homoscedastic variance of the error term of the second stage (scalar)}

\item{n.instruments}{number of instruments used in the analysis. For convenience,
the last \code{n.instruments} of the first stage design matrix are interpreted
as instruments. Hence, \code{n.instruments} cannot be larger than the number
of predictors in the first stage model, i.e. d1 (scalar)}
}
\description{
This function simulates panel data with time-varying parameters for treatment
effect analysis using instrumental variables. The (endogenous) treatment
has to be binary and a Gaussian observation equation is considered. Data
are then simulated as for non-treatment-effects models with the function \code{\link[=sim_panelTVP]{sim_panelTVP()}}
}
\examples{
sim.iv <- sim_panelTVP_IV(n = 1000,
                          Tmax = 4,
                          beta_stage1 = c(1, 0.7),
                          theta_stage1 = c(0.2, 0.01),
                          beta_stage2 = c(0, 4, 1),
                          theta_stage2 = c(0, 3, 0),
                          lambda_stage2 = 1.3,
                          psi_stage2 = 0.1,
                          beta_D = 2,
                          theta_D = 0.7,
                          rho = 0.1,
                          sigma2 = 1,
                          n.instruments = 1)
}
