% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_panelTVP_ZINB.R
\name{sim_panelTVP_ZINB}
\alias{sim_panelTVP_ZINB}
\title{Simulate data from a Zero-Inflated Negative Binomial (ZINB) panel with
time-varying parameters.}
\usage{
sim_panelTVP_ZINB(
  n,
  Tmax,
  beta.nb,
  theta.nb,
  lambda.nb,
  psi.nb,
  beta.logit,
  theta.logit,
  lambda.logit,
  psi.logit,
  r
)
}
\arguments{
\item{n}{number of subjects (scalar)}

\item{Tmax}{number of time points / repeated measurements per subject (scalar)}

\item{beta.nb}{fixed regression effects, with the first value representing the
global intercept (vector of dimension d_nb) (count model)}

\item{theta.nb}{standard deviation of random walk for regression effects, i.e.,
larger values yield regression effects that vary stronger over time
(vector of dimension d_nb) (count model)}

\item{lambda.nb}{fixed factor loading (scalar) (count model)}

\item{psi.nb}{standard deviation of random walk for factor loading, i.e.,
a larger value yields a factor loading that varies stronger over time (scalar)
(count model)}

\item{beta.logit}{fixed regression effects, with the first value representing the
global intercept (vector of dimension d_logit) (zero-inflation model)}

\item{theta.logit}{standard deviation of random walk for regression effects, i.e.,
larger values yield regression effects that vary stronger over time
(vector of dimension d_logit) (zero-inflation model)}

\item{lambda.logit}{fixed factor loading (scalar) (zero-inflation model)}

\item{psi.logit}{standard deviation of random walk for factor loading, i.e.,
a larger value yields a factor loading that varies stronger over time (scalar)
(zero-inflation model)}

\item{r}{dispersion parameter of the Negative Binomial component of the ZINB model}
}
\value{
a list of simulated data and parameters that contains the following elements:
\describe{
\item{observed}{a data frame that contains the response variable
\code{y} the covariates of the zero-inflation component
(starting with the letter \code{W} and ending with \code{logit}), the
covariates of the count component (starting with the letter \code{W} and
ending with \code{nb}), the time index \code{t} and the subject index \code{id}}
\item{beta.logit}{a T x d_logit matrix of regression effects
for the zero-inflation component of the model, i.e., each row contains
the regression effects of the corresponding time point}
\item{beta.nb}{a T x d_nb matrix of regression effects
for the count component of the model, i.e., each row contains
the regression effects of the corresponding time point}
\item{lambda.logit}{a T x 1 matrix of factor loadings for the
zero-inflation component of the model, i.e., each row contains
the factor loading of the corresponding time point}
\item{lambda.nb}{a T x 1 matrix of factor loadings for the
count component of the model, i.e., each row contains
the factor loading of the corresponding time point}
\item{r}{a scalar containing the value of \eqn{r} for the count component of the model}
}
}
\description{
Simulate data from a Zero-Inflated Negative Binomial (ZINB) panel with
time-varying parameters.
}
\examples{
x <- sim_panelTVP_ZINB(n = 100,
                       Tmax = 4,
                       beta.nb = c(0.5,0,-0.2),
                       theta.nb = c(0.1,0,0),
                       lambda.nb = 0.3,
                       psi.nb = 0,
                       beta.logit = c(0.9,0.1,0),
                       theta.logit = c(0.05,0,0),
                       lambda.logit = 0.8,
                       psi.logit = 0.1,
                       r = 2)
head(x$observed, 10)
x$beta.logit
x$beta.nb
x$lambda.logit
x$lambda.nb
x$r
}
