% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_summary.R
\name{summary.panelTVP.IV}
\alias{summary.panelTVP.IV}
\title{Summary output for a \code{panelTVP.IV} object}
\usage{
\method{summary}{panelTVP.IV}(object, by = "covariate", ...)
}
\arguments{
\item{object}{an object of class \code{panelTVP.IV}}

\item{by}{a single character that is either 'timepoint' or 'covariate'
to sort the output either by time point or covariate, respectively (default is covariate)}

\item{...}{optional arguments passed to the function (those are ignored)}
}
\description{
This \code{summary} function prints out a table that contains HPD-intervals,
posterior means and standard deviations based on the posterior distributions of the
time-varying parameters in a Bayesian IV model.
The results are sorted either by covariate (default) or by time point.
}
\examples{
# Computing summary table for object of class panelTVP.IV
# NB: To reduce computational effort, we have drastically reduced the length
# of the Markov Chain. You should use a much longer chain in your applications.
sim.iv <- sim_panelTVP_IV(n = 1000,
                          Tmax = 4,
                          beta_stage1 = c(1, 0.7),
                          theta_stage1 = c(0.2, 0.01),
                          beta_stage2 = c(0, 4, 1),
                          theta_stage2 = c(0, 3, 0),
                          lambda_stage2 = 1.3,
                          psi_stage2 = 0.1,
                          beta_D = 2,
                          theta_D = 0.7,
                          rho = 0.1,
                          sigma2 = 1,
                          n.instruments = 1)
res.iv <- panelTVP_IV(formula_stage1 = D ~ X_stage1.Z1,
                      formula_stage2 = y ~ X_stage2.W1 + X_stage2.W2 + D,
                      data = sim.iv$observed,
                      id = sim.iv$observed$id,
                      t = sim.iv$observed$t,
                      prior.rho = list(
                      mean.rho = atanh(0.1), sd.rho = 0.1,
                       expansion.steps = 10, width = 0.1
                      ),
                      mcmc.opt = list(chain.length = 200, burnin = 100, thin = 1, asis = TRUE))
summary(res.iv) # default = ordering by covariate
summary(res.iv, by = "timepoint") # ordering by time point
}
\author{
Roman Pfeiler, Helga Wagner
}
