% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_summary.R
\name{summary.panelTVP.NegBin}
\alias{summary.panelTVP.NegBin}
\title{Summary output for a Negative Binomial TVP model}
\usage{
\method{summary}{panelTVP.NegBin}(object, by = "covariate", ...)
}
\arguments{
\item{object}{an object of class \code{panelTVP.NegBin}}

\item{by}{a single character that is either 'timepoint' or 'covariate'
to sort the output either by time point or covariate, respectively (default is covariate)}

\item{...}{optional arguments passed to the function (those are ignored)}
}
\description{
This \code{summary} function prints out a table that contains HPD-intervals,
posterior means and standard deviations based on the posterior distributions of the
time-varying parameters, i.e., \eqn{\boldsymbol{\beta}_1,\dots,\boldsymbol{\beta}_T} and
\eqn{\lambda_1,\dots,\lambda_T}. The results are presented either sorted by covariate
(default) or by time point.
}
\author{
Roman Pfeiler
}
