% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_summary.R
\name{summary.panelTVP.ZINB}
\alias{summary.panelTVP.ZINB}
\title{Summary output for a \code{panelTVP.ZINB} object}
\usage{
\method{summary}{panelTVP.ZINB}(object, by = "covariate", ...)
}
\arguments{
\item{object}{an object of class \code{panelTVP.ZINB}}

\item{by}{a single character that is either 'timepoint' or 'covariate'
to sort the output either by time point or covariate, respectively (default is covariate)}

\item{...}{optional arguments passed to the function (those are ignored)}
}
\description{
This \code{summary} function prints out a table that contains HPD-intervals,
posterior means and standard deviations based on the posterior distributions of the
time-varying parameters, i.e., \eqn{\boldsymbol{\beta}_1,\dots,\boldsymbol{\beta}_T} and
\eqn{\lambda_1,\dots,\lambda_T}. The user can choose whether to show the results
of the Negative Binomial (count) model or the zer-inflation (Logit) model.
The results are presented either sorted by covariate (default) or by time point.
}
\examples{
# Computing summary table for object of class panelTVP.ZINB
# NB: To reduces computational effort, we have drastically reduced the length
# of the Markov Chain. You should use a much longer chain in your applications.
sim.zinb <- sim_panelTVP(n = 100,
                         Tmax = 4,
                         beta.nb = c(0.5,-0.7,0,0),
                         theta.nb = c(0.05,0.5,0,0),
                         lambda.nb = 0.5,
                         psi.nb = 0.02,
                         beta.logit = c(-1,0.6,0,0),
                         theta.logit = c(0,1,0,0),
                         lambda.logit = 0.7,
                         psi.logit = 0,
                         r = 2,
                         model = "ZINB")
res.zinb <- panelTVP(y ~ W1.nb + W2.nb + W3.nb | W1.logit + W2.logit + W3.logit,
                     data = sim.zinb$observed,
                     mcmc.opt = list(chain.length = 200, burnin = 100, thin = 1, asis = TRUE),
                     model = "ZINB")
# plots for count / Negative Binomial component
summary(res.zinb, component = "NegBin") # default = ordering by covariate
summary(res.zinb, by = "timepoint", component = "NegBin") # ordering by time point
# plots for zero-inflation / Logit component
summary(res.zinb, component = "Logit") # default = ordering by covariate
summary(res.zinb, by = "timepoint", component = "Logit") # ordering by time point
}
\author{
Roman Pfeiler
}
