% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelTVP.R
\name{panelTVP}
\alias{panelTVP}
\title{Fit a Bayesian panel model with time-varying parameters}
\usage{
panelTVP(
  formula,
  data,
  model,
  prior.reg = list(e1 = 0.001, e2 = 0.001, d1 = 0.001, d2 = 0.001, b_tau = 10, nu_tau =
    5, b_xi = 10, nu_xi = 5, a_tau = 1, kappa_tau = 10, a_xi = 1, kappa_xi = 10,
    iota.reg_tau = 1, iota.reg_xi = 1, learn_a_tau = TRUE, learn_a_xi = TRUE,
    tau.target.rate = 0.44, xi.target.rate = 0.44, learn_kappa_tau = TRUE, learn_kappa_xi
    = TRUE, type = "rw2", c = 1, B0 = 1e+05),
  prior.var = list(learn.C0.hyp = list(g0 = 5, G0 = 3.333333), c0 = 2.5),
  prior.load = list(e1 = 0.001, e2 = 0.001, d1 = 0.001, d2 = 0.001, b_phi = 10, nu_phi =
    5, b_zeta = 10, nu_zeta = 5, a_phi = 1, kappa_phi = 10, a_zeta = 1, kappa_zeta = 10,
    learn_kappa_phi = TRUE, learn_kappa_zeta = TRUE, type = "rw2", c = 1, L0 = 1),
  mcmc.opt = list(chain.length = 12000, burnin = 2000, thin = 10, asis = TRUE),
  settings.NegBin = list(iota.r = 0.1581139, r.a = 0.1, r.b = 10, r.target.rate = 0.44,
    slice = TRUE, r.alpha = 2, r.beta = 1, expansion.steps = 20, width = 1, p.overrelax =
    0, accuracy.overrelax = 20),
  HPD.coverage = 0.95
)
}
\arguments{
\item{formula}{The usual formula argument in regression methods, e.g., as in \code{\link[=lm]{lm()}}
(no default)}

\item{data}{A data.frame containing the variables specified in the formula argument
(no default)}

\item{model}{A character indicating which model you want to estimate
This parameter is either 'Gaussian', 'Probit', 'Logit' or 'NegBin' depending on
whether you want to fit a model for Gaussian, Probit, Logit or Negative Binomial
response data. In case you want to estimate a Zero-Inflated Negative Binomial
model, please use the function \code{\link[=panelTVP_ZINB]{panelTVP_ZINB()}}, which was designed for exactly
this situation}

\item{prior.reg}{A list of arguments for estimating the parameters of the regression
part of the model (see details)}

\item{prior.var}{A list of arguments for estimating the homoscedastic error variance
in a Gaussian/Normal model. For other models, this argument is ignored (see details)}

\item{prior.load}{A list of arguments for estimating the parameters of the factor
part of the model (see details)}

\item{mcmc.opt}{A list containing information for the overall sampler (see details)}

\item{settings.NegBin}{A list containing information for sampling the dispersion
parameter r in the Negative Binomial model. For other response distributions,
this argument is ignored (see details)}

\item{HPD.coverage}{Coverage probability of highest posterior density intervals
(default yields 95 percent coverage)}
}
\value{
The function returns an object of class \code{"panelTVP.Gaussian"},
\code{"panelTVP.Probit"}, \code{"panelTVP.Logit"} or \code{"panelTVP.NegBin"}
depending on whether a Gaussian, Probit, Logit or Negative Binomial model was
fitted. The returned object contains a list of the following elements
\describe{
\item{data}{the data used for fitting the model and additional context information
derived from the data}
\item{mcmc}{Markov Chains for every parameter except for the factor scores
(to save memory)}
\item{posterior}{preliminary summary of posterior results}
\item{fmean}{posterior means of random effects}
\item{model}{the fitted model}
\item{acceptance.rates}{the achieved acceptance rates when using Metropolis-Hastings}
\item{HPD.coverage}{coverage probability of HPD intervals (based on input)}
\item{runtime}{total runtime of the sampler (measured in seconds)}
\item{WAIC}{the Widely Applicable Information Criterion for model comparison}
\item{learning.settings}{information on which parameters have been learned}
\item{mcmc.settings}{details on general MCMC sampler}
}
}
\description{
This is the main function for fitting a flexible Bayesian regression model for
time-varying parameters in a balanced panel, i.e., when a total of n subjects are observed
at the same T time points. By using shrinkage priors, the model makes it
possible to identify whether an effect is time-varying, time-invariant or zero.
This function works for Gaussian, binary and Negative Binomial response data.
}
\examples{
1
}
