% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelTVP_ZINB.R
\name{panelTVP_ZINB}
\alias{panelTVP_ZINB}
\title{Fit a Bayesian panel model with time-varying parameters for overdispersed and zero-inflated count data}
\usage{
panelTVP_ZINB(
  formula,
  data,
  prior.reg_nb = list(e1 = 0.001, e2 = 0.001, d1 = 0.001, d2 = 0.001, b_tau = 10, nu_tau
    = 5, b_xi = 10, nu_xi = 5, a_tau = 1, kappa_tau = 10, a_xi = 1, kappa_xi = 10,
    iota.reg_tau = 1, iota.reg_xi = 1, learn_a_tau = TRUE, learn_a_xi = TRUE,
    tau.target.rate = 0.44, xi.target.rate = 0.44, learn_kappa_tau = TRUE, learn_kappa_xi
    = TRUE, type = "rw2", c = 1, B0 = 1e+05),
  prior.load_nb = list(e1 = 0.001, e2 = 0.001, d1 = 0.001, d2 = 0.001, b_phi = 10, nu_phi
    = 5, b_zeta = 10, nu_zeta = 5, a_phi = 1, kappa_phi = 10, a_zeta = 1, kappa_zeta =
    10, learn_kappa_phi = TRUE, learn_kappa_zeta = TRUE, type = "rw2", c = 1, L0 = 1),
  prior.reg_logit = list(e1 = 0.001, e2 = 0.001, d1 = 0.001, d2 = 0.001, b_tau = 10,
    nu_tau = 5, b_xi = 10, nu_xi = 5, a_tau = 1, kappa_tau = 10, a_xi = 1, kappa_xi = 10,
    iota.reg_tau = 1, iota.reg_xi = 1, learn_a_tau = TRUE, learn_a_xi = TRUE,
    tau.target.rate = 0.44, xi.target.rate = 0.44, learn_kappa_tau = TRUE, learn_kappa_xi
    = TRUE, type = "rw2", c = 1, B0 = 1e+05),
  prior.load_logit = list(e1 = 0.001, e2 = 0.001, d1 = 0.001, d2 = 0.001, b_phi = 10,
    nu_phi = 5, b_zeta = 10, nu_zeta = 5, a_phi = 1, kappa_phi = 10, a_zeta = 1,
    kappa_zeta = 10, learn_kappa_phi = TRUE, learn_kappa_zeta = TRUE, type = "rw2", c =
    1, L0 = 1),
  mcmc.opt = list(chain.length = 12000, thin = 10, burnin = 2000, asis = TRUE),
  settings.NegBin = list(iota.r = 0.1581139, r.a = 0.1, r.b = 10, r.target.rate = 0.44,
    slice = TRUE, r.alpha = 2, r.beta = 1, expansion.steps = 20, width = 1, p.overrelax =
    0, accuracy.overrelax = 20),
  HPD.coverage = 0.95
)
}
\arguments{
\item{formula}{A formula argument consisting of two parts: the count model and
the zero-inflation model (no default, see details)}

\item{data}{A data.frame containing the variables specified in the formula argument
(no default)}

\item{prior.reg_nb}{A list of arguments for estimating the parameters of the regression
part of the count model (see details)}

\item{prior.load_nb}{A list of arguments for estimating the parameters of the factor
part of the count model (see details)}

\item{prior.reg_logit}{A list of arguments for estimating the parameters of the regression
part of the zero-inflation model (see details)}

\item{prior.load_logit}{A list of arguments for estimating the parameters of the factor
part of the zero-inflation model (see details)}

\item{mcmc.opt}{A list containing information for the overall sampler (see details)}

\item{settings.NegBin}{A list containing information for sampling the dispersion
parameter r of the count model (see details)}

\item{HPD.coverage}{Coverage probability of highest posterior density intervals
(default yields 95 percent coverage)}
}
\value{
nix nix
}
\description{
Fit a Bayesian panel model with time-varying parameters for overdispersed and zero-inflated count data
}
\examples{
0
}
