% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_plot.R
\name{plot.panelTVP.Probit}
\alias{plot.panelTVP.Probit}
\title{Get plots of time-varying parameters for an object of class \code{panelTVP.Probit}}
\usage{
\method{plot}{panelTVP.Probit}(x, nplots = 4, ...)
}
\arguments{
\item{x}{an object of class \code{panelTVP.Probit}}

\item{nplots}{indicates how many plots should be printed on one page}

\item{...}{optional arguments passed to the function (those are ignored)}
}
\description{
This \code{plot} function produces effect plots for the time-varying parameters,
i.e., \eqn{\boldsymbol{\beta}_1,\dots,\boldsymbol{\beta}_T} and
\eqn{\lambda_1,\dots,\lambda_T} separately for each covariate Plots are produced
using the \code{ggplot2} library.
}
\author{
Roman Pfeiler
}
