% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_plot.R
\name{plot.panelTVP.ZINB}
\alias{plot.panelTVP.ZINB}
\title{Get plots of time-varying parameters for an object of class \code{panelTVP.ZINB}}
\usage{
\method{plot}{panelTVP.ZINB}(x, component = NULL, nplots = 4, ...)
}
\arguments{
\item{x}{an object of class \code{panelTVP.ZINB}}

\item{component}{either 'NegBin' or 'Logit' to create plots for either the count or the
zero-inflation component, respectively - must be specified!}

\item{nplots}{indicates how many plots should be printed on one page}

\item{...}{optional arguments passed to the function (those are ignored)}
}
\description{
This \code{plot} function produces effect plots for the time-varying parameters,
i.e., \eqn{\boldsymbol{\beta}_1,\dots,\boldsymbol{\beta}_T} and
\eqn{\lambda_1,\dots,\lambda_T} separately for each covariate. Plots are produced
using the \code{ggplot2} library. Note that this
function either plots the effects of the Negative Binomial (or count) model, or
the Logit (or zero-inflation) model.
}
\author{
Roman Pfeiler
}
