% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_plot.R
\name{plot.panelTVP.ZINB}
\alias{plot.panelTVP.ZINB}
\title{Get plots of time-varying parameters based on a \code{panelTVP.ZINB} object}
\usage{
\method{plot}{panelTVP.ZINB}(x, component = NULL, nplots = 4, ...)
}
\arguments{
\item{x}{an object of class \code{panelTVP.ZINB}}

\item{component}{either 'NegBin' or 'Logit' to create plots for either the count or the
zero-inflation component (no default)}

\item{nplots}{indicates how many plots should be printed on one page}

\item{...}{optional arguments passed to the function (those are ignored)}
}
\description{
This \code{plot} function produces effect plots for the time-varying parameters,
i.e., \eqn{\boldsymbol{\beta}_1,\dots,\boldsymbol{\beta}_T} and
\eqn{\lambda_1,\dots,\lambda_T} separately for each covariate. Plots are produced
using the \code{ggplot2} library. Note that this
function either plots the effects of the Negative Binomial (or count) model, or
the Logit (or zero-inflation) model.
}
\examples{
# Plot results from an object of class panelTVP.ZINB
# NB: To reduce computational effort, we have drastically reduced the length
# of the Markov Chain. You should use a much longer chain in your applications.
sim.zinb <- sim_panelTVP(n = 100,
                         Tmax = 4,
                         beta.nb = c(0.5,-0.7,0,0),
                         theta.nb = c(0.05,0.5,0,0),
                         lambda.nb = 0.5,
                         psi.nb = 0.02,
                         beta.logit = c(-1,0.6,0,0),
                         theta.logit = c(0,1,0,0),
                         lambda.logit = 0.7,
                         psi.logit = 0,
                         r = 2,
                         model = "ZINB")
res.zinb <- panelTVP(y ~ W1.nb + W2.nb + W3.nb | W1.logit + W2.logit + W3.logit,
                     data = sim.zinb$observed,
                     mcmc.opt = list(chain.length = 200, burnin = 100, thin = 1, asis = TRUE),
                     model = "ZINB")
plot(res.zinb, nplots = 1, component = "NegBin") # effects that belong to count component
plot(res.zinb, nplots = 1, component = "Logit") # effects that belong to zero-inflation component
}
\references{
Wickham, H. (2016). \code{ggplot2}: Elegant Graphics for Data Analysis.
Springer Verlag, New York.
}
\author{
Roman Pfeiler, Helga Wagner
}
