% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_predict.R
\name{predict.panelTVP.Probit}
\alias{predict.panelTVP.Probit}
\title{Get predictions for new subjects}
\usage{
\method{predict}{panelTVP.Probit}(object, X.new, timepoint, coverage = 0.95, pop.pred = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{panelTVP.Probit}}

\item{X.new}{a matrix or data frame consisting of new data for the same variables
that were used for fitting the model.}

\item{timepoint}{a numeric scalar indicating the time point for which predictions should
be made, i.e., predictions for a given data set are only made for one specific time point.
In case you want predictions for additional time points you need to repeatedly call this
function.}

\item{coverage}{coverage probability for prediction intervals - defaults to 95 \% coverage}

\item{pop.pred}{logical value, if TRUE population-based predictions are made,
that ignore the random effects structure, if FALSE the random effects structure
is included as well where the unknown factor scores are sampled from their
standard Normal prior - defaults to FALSE}

\item{...}{optional arguments passed to the function (those are ignored)}
}
\value{
a list containing the following elements
\itemize{
\item \code{predictive.distribution}: a matrix where each row contains draws
from the posterior predictive distribution for the corresponding observation
\item \code{predictive.summary}: posterior mean as well as HPD interval based on
posterior predictive distribution, i.e., the most basic summary statistics
}
}
\description{
This \code{predict} function simulates data and computes summary statistics based on the
posterior predictive distribution for new data under a Bernoulli likelihood using a
Probit link.
}
\author{
Roman Pfeiler
}
