% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_summary.R
\name{summary.panelTVP.ZINB}
\alias{summary.panelTVP.ZINB}
\title{Summary output for a \code{panelTVP.ZINB} object}
\usage{
\method{summary}{panelTVP.ZINB}(object, by = "covariate", ...)
}
\arguments{
\item{object}{an object of class \code{panelTVP.ZINB}}

\item{by}{a single character that is either 'timepoint' or 'covariate'
to sort the output either by time point or covariate, respectively (default is covariate)}

\item{...}{optional arguments passed to the function (those are ignored)}
}
\description{
This \code{summary} function prints out a table that contains HPD-intervals,
posterior means and standard deviations based on the posterior distributions of the
time-varying parameters, i.e., \eqn{\boldsymbol{\beta}_1,\dots,\boldsymbol{\beta}_T} and
\eqn{\lambda_1,\dots,\lambda_T}. The results are sorted either by covariate (default)
or by time point.
}
\examples{
# Computing summary table for object of class panelTVP.ZINB
# NB: To reduce computational effort, we have drastically reduced the length
# of the Markov Chain. You should use a much longer chain in your applications.
sim.zinb <- sim_panelTVP(n = 100,
                         Tmax = 4,
                         beta_zinb.count = c(0.5,-0.7,0,0),
                         theta_zinb.count = c(0.05,0.5,0,0),
                         lambda_zinb.count = 0.5,
                         psi_zinb.count = 0.02,
                         beta_zinb.inflation = c(-1,0.6,0,0),
                         theta_zinb.inflation = c(0,1,0,0),
                         lambda_zinb.inflation = 0.7,
                         psi_zinb.inflation = 0,
                         r = 2,
                         model = "ZINB")
res.zinb <- panelTVP(y ~ W1.nb + W2.nb + W3.nb | W1.logit + W2.logit + W3.logit,
                     data = sim.zinb$observed,
                     id = sim.zinb$observed$id,
                     t = sim.zinb$observed$t,
                     mcmc.opt = list(chain.length = 200, burnin = 100, thin = 1, asis = TRUE),
                     model = "ZINB")
summary(res.zinb) # default = ordering by covariate
summary(res.zinb, by = "timepoint") # ordering by time point
}
\author{
Roman Pfeiler, Helga Wagner
}
